+++ examples/resources/iosxe_vlan/import.sh
@@ -1,0 +1,1 @@
+terraform import iosxe_vlan.example "Cisco-IOS-XE-native:native/vlan/ios-vlan:vlan-list=123"