+++ examples/resources/iosxe_vlan_configuration/import.sh
@@ -1,0 +1,1 @@
+terraform import iosxe_vlan_configuration.example "Cisco-IOS-XE-native:native/vlan/ios-vlan:configuration-entry=123"