+++ examples/resources/iosxe_vrf/resource.tf
@@ -7,12 +7,14 @@ resource "iosxe_vrf" "example" {
   vpn_id              = "1:1"
   route_target_import = [
     {
-      value = "1:1"
+      value     = "1:1"
+      stitching = false
     }
   ]
   route_target_export = [
     {
-      value = "1:1"
+      value     = "1:1"
+      stitching = false
     }
   ]
 }