+++ internal/provider/resource_iosxe_vrf_test.go
@@ -23,7 +23,9 @@ func TestAccIosxeVRF(t *testing.T) {
 					resource.TestCheckResourceAttr("iosxe_vrf.test", "address_family_ipv6", "true"),
 					resource.TestCheckResourceAttr("iosxe_vrf.test", "vpn_id", "1:1"),
 					resource.TestCheckResourceAttr("iosxe_vrf.test", "route_target_import.0.value", "1:1"),
+					resource.TestCheckResourceAttr("iosxe_vrf.test", "route_target_import.0.stitching", "false"),
 					resource.TestCheckResourceAttr("iosxe_vrf.test", "route_target_export.0.value", "1:1"),
+					resource.TestCheckResourceAttr("iosxe_vrf.test", "route_target_export.0.stitching", "false"),
 				),
 			},
 			{
@@ -54,9 +56,11 @@ func testAccIosxeVRFConfig_all() string {
 		vpn_id = "1:1"
 		route_target_import = [{
 		value = "1:1"
+		stitching = false
 		}]
 		route_target_export = [{
 		value = "1:1"
+		stitching = false
 		}]
 	}
 	`