+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.1.1 (unreleased)
+
+- Add iosxe_vlan resource and data source
+- Add iosxe_vlan_configuration resource and data source
+- Add stitching option to vrf route targets
+
 ## 0.1.0
 
 - Initial release